stage1_analysis/source_finding/requirements.txt
# Source Finding Module Requirements

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# LLM APIs
openai>=1.0.0

# Environment management
python-dotenv>=1.0.0

# Progress bars
tqdm>=4.65.0

# Optional: for notebook
jupyter>=1.0.0
ipykernel>=6.0.0

